+++ src/Akeneo/Apps/back/Infrastructure/Symfony/Resources/public/js/controller/apps.tsx
@@ -2,7 +2,7 @@ import * as $ from 'jquery';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import {Apps, RouterProvider} from '@akeneo-pim-ce/apps';
+import {Index, RouterProvider} from '@akeneo-pim-ce/apps';
 
 const BaseController = require('pim/controller/base');
 const Mediator = require('oro/mediator');
@@ -14,14 +14,14 @@ const provideServices = (Component: any) => (props: any) => (
   </RouterProvider>
 );
 
-const AppsWithServices = provideServices(Apps);
+const IndexWithServices = provideServices(Index);
 
 class AppsController extends BaseController {
   renderRoute() {
     Mediator.trigger('pim_menu:highlight:tab', {extension: 'pim-menu-system'});
     Mediator.trigger('pim_menu:highlight:item', {extension: 'pim-menu-system-apps'});
 
-    ReactDOM.render(<AppsWithServices />, this.el);
+    ReactDOM.render(<IndexWithServices />, this.el);
 
     return $.Deferred().resolve();
   }
