+++ src/Akeneo/Apps/front/src/__tests__/application/common/button/Button.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {create} from 'react-test-renderer';
-import {Button} from '../../../../application/common/button/button.component';
+import {Button} from '../../../../application/common/button/Button';
 
 describe('Button component', () => {
     test('Matches the snapshot', () => {
