+++ src/Akeneo/Apps/front/src/__tests__/application/shared/translate/Translate.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {create} from 'react-test-renderer';
 import {TranslateContext} from '../../../../application/shared/translate/translate-context';
-import {Translate} from '../../../../application/shared/translate/translate.component';
+import {Translate} from '../../../../application/shared/translate/Translate';
 
 describe('Translate component', () => {
     test('Matches the snapshot', () => {
