--- src/Akeneo/Apps/front/src/application/app/edit-app.component.tsx
@@ -1,3 +1,0 @@
-import * as React from 'react';
-
-export const EditApp = () => <>Edit App</>;
