+++ src/Akeneo/Apps/front/src/application/apps/components/AppForm.tsx
@@ -1,0 +1,3 @@
+import * as React from 'react';
+
+export const AppForm = () => <>AppForm</>;
