+++ src/Akeneo/Apps/front/src/application/apps/pages/EditApp.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import {Page} from '../../common/Page';
+import {AppForm} from '../components/AppForm';
+
+export const EditApp = () => (
+    <Page>
+        EditApp
+        <AppForm />
+    </Page>
+);
