+++ src/Akeneo/Apps/front/src/application/apps/pages/Index.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import {HashRouter as Router, Route, Switch} from 'react-router-dom';
+import {EditApp} from './EditApp';
+import {ListApp} from './ListApp';
+
+export const Index = () => (
+    <Router>
+        <Switch>
+            <Route path='/apps/edit'>
+                <EditApp />
+            </Route>
+            <Route path='/apps'>
+                <ListApp />
+            </Route>
+        </Switch>
+    </Router>
+);
