+++ src/Akeneo/Apps/front/src/application/common/Page.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 
-import {PimView} from '../../infrastructure/pim-view.component';
+import {PimView} from '../../infrastructure/PimView';
 
 type Props = React.PropsWithChildren<{
     breadcrumb?: any;
-    actionButtons?: any[];
-    title?: any
+    buttons?: React.ReactElement[];
+    title?: any;
 }>;
 
-export const Page = ({children, breadcrumb, actionButtons, title}: Props) => (
+export const Page = ({children, breadcrumb, buttons, title}: Props) => (
     <div className='AknDefault-contentWithColumn'>
         <div className='AknDefault-thirdColumnContainer'>
             <div className='AknDefault-thirdColumn' />
@@ -26,9 +26,11 @@ export const Page = ({children, breadcrumb, actionButtons, title}: Props) => (
                                         className='AknTitleContainer-userMenuContainer AknTitleContainer-userMenu'
                                         viewName='pim-apps-user-navigation'
                                     />
-                                    {actionButtons && (
+                                    {buttons && (
                                         <div className='AknTitleContainer-actionsContainer AknButtonList'>
-                                            {actionButtons}
+                                            {buttons.map((button, index) => (
+                                                <React.Fragment key={index}>{button}</React.Fragment>
+                                            ))}
                                         </div>
                                     )}
                                 </div>
