+++ src/Akeneo/Apps/front/src/application/common/button/ActionButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {PropsWithChildren} from 'react';
-import { Button, Props } from './button.component';
+import { Button, Props } from './Button';
 
 export const ActionButton = ({classNames = [], ...props}: PropsWithChildren<Props>) => {
     classNames.push('AknButton--action');
