+++ src/Akeneo/Apps/front/src/application/shared/router/with-router.hoc-factory.tsx
