+++ src/Akeneo/Apps/front/src/application/shared/translate/Translate.tsx
