+++ src/Akeneo/Apps/front/src/index.ts
@@ -1,6 +1,6 @@
-import {ListApp} from './application/app/list-app.component';
+import {Index} from './application/apps/pages/Index';
 import {RouterContext} from './application/shared/router/router-context';
 
 const RouterProvider = RouterContext.Provider;
 
-export {ListApp as Apps, RouterProvider};
+export {Index, RouterProvider};
