+++ src/Akeneo/Apps/front/src/infrastructure/PimView.tsx
